+++ app/src/main/java/com/example/gradeprojectdb01/data/entities/TunSysParameter.kt
@@ -14,8 +14,8 @@ import com.example.gradeprojectdb01.data.convertersAndEnums.ParamValType
 
 data class TunSysParameter (
     @PrimaryKey (autoGenerate = true)
-    val tunSysParamId: Int,
-    val tunSysId: Int,
+    val tunSysParamId: Long,
+    val tunSysId: Long,
     val valueType : ParamValType,
     val valueName: String,
     val value: String,
